requirements.txt
PyYAML~=5.3.1
xlrd~=1.2.0
requests~=2.24.0
jsonpath~=0.82
pytest~=6.0.1